requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.58.2
django-appconf==1.0.5
django-jwt-auth-middleware==0.1.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.1.3
factory-boy==3.3.0
Faker==20.0.3
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
oauthlib==3.2.2
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==1.7.1
pytest==7.4.3
pytest-django==4.7.0
pytest-factoryboy==2.6.0
python-dateutil==2.8.2
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0